Fix plugin restart issue (probably a crash)
Even though `acc.AddFields` accepts a map[string]interface{}, storing
arbitrary structs in that map leads to the plugin crashing (possibly because
it doesn't know how to serialize it?).

This commit reduces the data sent to a single float64, which stabilizes the
plugin, but possibly loses data if more than 1 data point is returned in a
response. Immediately following this fix we should perform some data
munging to make sure that all data is collected.